requirements.txt
fastapi
uvicorn
SQLAlchemy
pymysql
python-dotenv
celery
redis
Pillow